+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 6.3.0
+
+* Fix and reintroduce Add item
+* Show which Bitwarden server is being used in the Login to Bitwarden title
+* Fix server URL for bitwarden.eu again
+
 ## 6.2.2
 
 * Fix server URL for bitwarden.eu